+++ okdata/cli/__main__.py
@@ -11,7 +11,7 @@ from okdata.cli.commands.permissions import PermissionsCommand
 from okdata.cli.commands.pipelines import Pipelines
 from okdata.cli.commands.pubreg import PubregCommand
 from okdata.cli.commands.status import StatusCommand
-from okdata.cli.commands.teams import TeamsCommand
+from okdata.cli.commands.teams.teams import TeamsCommand
 
 
 def main():